all new tool logic + removed unnecessary files

+++ macOS/writing-tools/UpdateChecker.swift
@@ -4,7 +4,7 @@ import AppKit
 @Observable
 final class UpdateChecker {
     static let shared = UpdateChecker()
-    private let currentVersion = 3  // Current app version
+    private let currentVersion = 3.0 // Current app version
     private let updateCheckURL = "https://raw.githubusercontent.com/theJayTea/WritingTools/main/macOS/Latest_Version_for_Update_Check.txt"
     private let updateDownloadURL = "https://github.com/theJayTea/WritingTools/releases"
     
@@ -46,7 +46,7 @@ final class UpdateChecker {
             
             if let versionString = cleanedString,
                !versionString.isEmpty,
-               let latestVersion = Int(versionString) {
+               let latestVersion = Double(versionString) {  // Changed to Double
                 print("Parsed version: \(latestVersion)")
                 updateAvailable = latestVersion > currentVersion
             } else {
